@@ -1,4 +1,3 @@
 class Shop < ActiveRecord::Base
-  include ShopifyApp::Shop
-  include ShopifyApp::SessionStorage
+  include SpiffyStoresApp::SessionStorage
 end
