@@ -2,7 +2,7 @@ require 'spiffy_stores_app/version'
 
 # deps
 require 'spiffy_stores_api'
-#require 'omniauth-shopify-oauth2'
+require 'omniauth-spiffy-oauth2'
 
 # config
 require 'spiffy_stores_app/configuration'
@@ -10,21 +10,25 @@ require 'spiffy_stores_app/configuration'
 # engine
 require 'spiffy_stores_app/engine'
 
-# jobs
-require 'spiffy_stores_app/webhooks_manager_job'
-require 'spiffy_stores_app/scripttags_manager_job'
-
-# helpers and concerns
-require 'spiffy_stores_app/shop'
-require 'spiffy_stores_app/session_storage'
-require 'spiffy_stores_app/sessions_concern'
-require 'spiffy_stores_app/login_protection'
-require 'spiffy_stores_app/webhooks_manager'
-require 'spiffy_stores_app/scripttags_manager'
-require 'spiffy_stores_app/webhook_verification'
-require 'spiffy_stores_app/app_proxy_verification'
+# utils
 require 'spiffy_stores_app/utils'
 
-# session repository
-require 'spiffy_stores_app/spiffy_stores_session_repository'
-require 'spiffy_stores_app/in_memory_session_store'
+# controller concerns
+require 'spiffy_stores_app/controller_concerns/localization'
+require 'spiffy_stores_app/controller_concerns/login_protection'
+require 'spiffy_stores_app/controller_concerns/embedded_app'
+require 'spiffy_stores_app/controller_concerns/webhook_verification'
+require 'spiffy_stores_app/controller_concerns/app_proxy_verification'
+
+# jobs
+require 'spiffy_stores_app/jobs/webhooks_manager_job'
+require 'spiffy_stores_app/jobs/scripttags_manager_job'
+
+# mangers
+require 'spiffy_stores_app/managers/webhooks_manager'
+require 'spiffy_stores_app/managers/scripttags_manager'
+
+# session
+require 'spiffy_stores_app/session/session_storage'
+require 'spiffy_stores_app/session/session_repository'
+require 'spiffy_stores_app/session/in_memory_session_store'
