@@ -1,12 +1,12 @@
 require 'test_helper'
-require 'generators/shopify_app/home_controller/home_controller_generator'
+require 'generators/spiffy_stores_app/home_controller/home_controller_generator'
 
 class HomeControllerGeneratorTest < Rails::Generators::TestCase
-  tests ShopifyApp::Generators::HomeControllerGenerator
+  tests SpiffyStoresApp::Generators::HomeControllerGenerator
   destination File.expand_path("../tmp", File.dirname(__FILE__))
 
   setup do
-    ShopifyApp.configure do |config|
+    SpiffyStoresApp.configure do |config|
       config.embedded_app = true
     end
 
@@ -24,27 +24,23 @@ class HomeControllerGeneratorTest < Rails::Generators::TestCase
   test "creates the home index view with embedded options" do
     run_generator
     assert_file "app/views/home/index.html.erb" do |index|
-      assert_match "ShopifyApp.ready", index
+      assert_match "SpiffyStoresApp.ready", index
     end
   end
 
   test "creates the home index view with embedded false" do
-    stub_embedded_false
+    SpiffyStoresApp.configuration.embedded_app = false
     run_generator
     assert_file "app/views/home/index.html.erb" do |index|
-      refute_match "ShopifyApp.ready", index
+      refute_match "SpiffyStoresApp.ready", index
     end
   end
 
   test "adds home route to routes" do
     run_generator
     assert_file "config/routes.rb" do |routes|
-      assert_match "mount ShopifyApp::Engine, at: '/'", routes
+      assert_match "mount SpiffyStoresApp::Engine, at: '/'", routes
       assert_match "root :to => 'home#index'", routes
     end
-  end
-
-  def stub_embedded_false
-    ShopifyApp.configuration.embedded_app = false
   end
 end
