@@ -1,8 +1,8 @@
 require 'test_helper'
-require 'generators/shopify_app/routes/routes_generator'
+require 'generators/spiffy_stores_app/routes/routes_generator'
 
 class ControllerGeneratorTest < Rails::Generators::TestCase
-  tests ShopifyApp::Generators::RoutesGenerator
+  tests SpiffyStoresApp::Generators::RoutesGenerator
   destination File.expand_path("../tmp", File.dirname(__FILE__))
 
   setup do
@@ -10,15 +10,15 @@ class ControllerGeneratorTest < Rails::Generators::TestCase
     provide_existing_routes_file
   end
 
-  test "copies ShopifyApp routes to the host application" do
+  test "copies SpiffyStoresApp routes to the host application" do
     run_generator
 
     assert_file "config/routes.rb" do |routes|
       assert_match "get 'login' => :new, :as => :login", routes
       assert_match "post 'login' => :create, :as => :authenticate", routes
-      assert_match "get 'auth/shopify/callback' => :callback", routes
+      assert_match "get 'auth/spiffy/callback' => :callback", routes
       assert_match "get 'logout' => :destroy, :as => :logout", routes
-      refute_match "mount ShopifyApp::Engine, at: '/'", routes
+      refute_match "mount SpiffyStoresApp::Engine, at: '/'", routes
     end
   end
 
