@@ -1,17 +1,20 @@
 require 'test_helper'
-require 'generators/shopify_app/shop_model/shop_model_generator'
+require 'generators/spiffy_stores_app/shop_model/shop_model_generator'
 
 class ShopModelGeneratorTest < Rails::Generators::TestCase
-  tests ShopifyApp::Generators::ShopModelGenerator
+  tests SpiffyStoresApp::Generators::ShopModelGenerator
   destination File.expand_path("../tmp", File.dirname(__FILE__))
-  setup :prepare_destination
+  
+  setup do
+    prepare_destination
+    provide_existing_initializer_file
+  end
 
   test "create the shop model" do
     run_generator
     assert_file "app/models/shop.rb" do |shop|
       assert_match "class Shop < ActiveRecord::Base", shop
-      assert_match "include ShopifyApp::Shop", shop
-      assert_match "include ShopifyApp::SessionStorage", shop
+      assert_match "include SpiffyStoresApp::SessionStorage", shop
     end
   end
 
@@ -22,10 +25,10 @@ class ShopModelGeneratorTest < Rails::Generators::TestCase
     end
   end
 
-  test "adds the shopify_session_repository initializer" do
+  test "updates the spiffy_stores_app initializer" do
     run_generator
-    assert_file "config/initializers/shopify_session_repository.rb" do |file|
-      assert_match "ShopifyApp::SessionRepository.storage = Shop", file
+    assert_file "config/initializers/spiffy_stores_app.rb" do |file|
+      assert_match "config.session_repository = Shop", file
     end
   end
 
