@@ -3,47 +3,47 @@ require 'test_helper'
 class ConfigurationTest < ActiveSupport::TestCase
 
   setup do
-    ShopifyApp.configuration = nil
+    SpiffyStoresApp.configuration = nil
   end
 
   test "configure" do
-    ShopifyApp.configure do |config|
+    SpiffyStoresApp.configure do |config|
       config.embedded_app = true
     end
 
-    assert_equal true, ShopifyApp.configuration.embedded_app
+    assert_equal true, SpiffyStoresApp.configuration.embedded_app
   end
 
-  test "defaults to myshopify_domain" do
-    assert_equal "myshopify.com", ShopifyApp.configuration.myshopify_domain
+  test "defaults to spiffy_stores_domain" do
+    assert_equal "spiffystores.com", SpiffyStoresApp.configuration.spiffy_stores_domain
   end
 
-  test "can set myshopify_domain" do
-    ShopifyApp.configure do |config|
-      config.myshopify_domain = 'myshopify.io'
+  test "can set spiffy_stores_domain" do
+    SpiffyStoresApp.configure do |config|
+      config.spiffy_stores_domain = 'spiffystores.io'
     end
 
-    assert_equal "myshopify.io", ShopifyApp.configuration.myshopify_domain
+    assert_equal "spiffystores.io", SpiffyStoresApp.configuration.spiffy_stores_domain
   end
 
   test "can configure webhooks for creation" do
     webhook = {topic: 'carts/update', address: 'example-app.com/webhooks', format: 'json'}
 
-    ShopifyApp.configure do |config|
+    SpiffyStoresApp.configure do |config|
       config.webhooks = [webhook]
     end
 
-    assert_equal webhook, ShopifyApp.configuration.webhooks.first
+    assert_equal webhook, SpiffyStoresApp.configuration.webhooks.first
   end
 
   test "has_webhooks? is true if webhooks have been configured" do
-    refute ShopifyApp.configuration.has_webhooks?
+    refute SpiffyStoresApp.configuration.has_webhooks?
 
-    ShopifyApp.configure do |config|
+    SpiffyStoresApp.configure do |config|
       config.webhooks = [{topic: 'carts/update', address: 'example-app.com/webhooks', format: 'json'}]
     end
 
-    assert ShopifyApp.configuration.has_webhooks?
+    assert SpiffyStoresApp.configuration.has_webhooks?
   end
 
 end
